Create callDataFromHDAP.py
This pulls call data from the Vonage Business platform in JSON format. Then it pulls out the names, number of calls, and total talk time for the specified reps. Finally, these lists are combined into a list of lists, so that this final list can be easily appended to an excel sheet.