Create sdrMorningAndAfternoon-OB
Running file via Windows Task Manager in once in the morning and once at night to run this script. It pulls data from Salesforce API and internal API. Puts this into existing Excel spreadsheets with formatting and formulas and then saves it in a folder for that month and year. Finally, it e-mails the designated supervisors or reps.